+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## v0.1.2 (dev)
 
+### Enhancements
+
   * Add some guards to functions
+
+### Bug fixes
+
+  * `A.RBSet.disjoint?/2` was not returning the expected value
 
 ## v0.1.1 (2020-10-21)